+++ TESTCASES/example_tc1_copy.py
@@ -1,5 +1,3 @@
-import time
-from testplans import TestCaseBase
 from . import example_tc1_direct