+++ _share-1=prepare-1=release_files.py
@@ -1,4 +1,0 @@
-from PROJECT import PROJECT
-from _aux__release_files import release_files_update
-
-release_files_update(PROJECT)