+++ testplans/__init__.py
@@ -1,6 +1,94 @@
-from .devices import *
-from .tc import *
-from .main import *
+# =====================================================================================================================
+# VERSION = (0, 0, 1)   # use import EXACT_OBJECTS! not *
+#   from .main import *                 # INcorrect
+#   from .main import EXACT_OBJECTS     # CORRECT
 
-from .gui import *
-from .tm import *
+
+# =====================================================================================================================
+# TEMPLATE
+# from .main import (
+#     # BASE
+#     EXACT_OBJECTS,
+#
+#     # AUX
+#
+#     # TYPES
+#
+#     # EXX
+# )
+# ---------------------------------------------------------------------------------------------------------------------
+from .main import (
+    # BASE
+    TpInsideApi_Runner,
+    TpMultyDutBase,
+
+    # AUX
+
+    # TYPES
+
+    # EXX
+    Exx__TcsPathNotExists,
+    Exx__TcItemNotFound,
+    Exx__TcItemType,
+    Exx__TcSettingsIncorrect,
+)
+from .devices import (
+    # BASE
+    DeviceBase,
+    DutBase,
+    DevicesBreeder,
+    DevicesBreeder_WithDut,
+
+    # AUX
+    DevicesBreeder_Example,
+
+    # TYPES
+
+    # EXX
+)
+from .tc import (
+    # BASE
+    TestCaseBase,
+
+    # AUX
+    Signals,
+
+    # TYPES
+    TYPE__RESULT,
+
+    # EXX
+)
+from .gui import (
+    # BASE
+    TpGuiBase,
+
+    # AUX
+
+    # TYPES
+
+    # EXX
+)
+from .tm import (
+    # BASE
+    TpTableModel,
+
+    # AUX
+
+    # TYPES
+
+    # EXX
+)
+from .api import (
+    # BASE
+    TpApi_Aiohttp,
+    TpApi_FastApi,
+    create_app__FastApi_Tp,
+
+    # AUX
+
+    # TYPES
+
+    # EXX
+)
+
+# =====================================================================================================================